@@ -1,21 +1,20 @@
 name: packagist_extensions
 description: extension tools
-version: 1.1.2
+version: 1.1.3
 homepage: https://github.com/flutter-packagist/extensions
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
-  flutter: ">=1.17.0"
+  sdk: '>=3.3.0 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  device_info_plus: ^10.1.0
-  get: ^4.6.6
+  device_info_plus: ^11.3.0
+  get: ^4.7.2
   intl: ^0.19.0
-  permission_handler: ^11.3.1
-  shared_preferences: ^2.2.3
-  synchronized: ^3.1.0+1
+  permission_handler: ^11.4.0
+  shared_preferences: ^2.5.2
+  synchronized: ^3.3.0
 
 dev_dependencies:
   flutter_test:
